+++ core-valid-adapter-app/src/main/java/com/farao_community/farao/core_valid/adapter/app/CoreValidAdapterApplication.java
@@ -1,0 +1,21 @@
+/*
+ * Copyright (c) 2021, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.farao_community.farao.core_valid.adapter.app;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * @author Alexandre Montigny {@literal <alexandre.montigny at rte-france.com>}
+ */
+@SuppressWarnings("hideutilityclassconstructor")
+@SpringBootApplication
+public class CoreValidAdapterApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(CoreValidAdapterApplication.class, args);
+    }
+}
