+++ core-valid-adapter-app/src/main/java/com/farao_community/farao/core_valid/adapter/app/CoreValidAdapterListener.java
@@ -1,11 +1,20 @@
-package com.farao_community.farao.corevalid.adapter;
+/*
+ * Copyright (c) 2021, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.farao_community.farao.core_valid.adapter.app;
 
 import com.farao_community.farao.core_valid.api.resource.CoreValidFileResource;
 import com.farao_community.farao.core_valid.api.resource.CoreValidRequest;
 import com.farao_community.farao.gridcapa.task_manager.api.ProcessFileDto;
 import com.farao_community.farao.gridcapa.task_manager.api.TaskDto;
 import com.farao_community.farao.gridcapa_core_valid.starter.CoreValidClient;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Bean;
+import org.springframework.stereotype.Component;
 
 import java.time.OffsetDateTime;
 import java.util.List;
@@ -14,8 +23,10 @@ import java.util.function.Consumer;
 /**
  * @author Alexandre Montigny {@literal <alexandre.montigny at rte-france.com>}
  */
+@Component
 public class CoreValidAdapterListener {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(CoreValidAdapterListener.class);
     private final CoreValidClient coreValidClient;
 
     public CoreValidAdapterListener(CoreValidClient coreValidClient) {
@@ -25,6 +36,7 @@ public class CoreValidAdapterListener {
     @Bean
     public Consumer<TaskDto> handleRun() {
         return taskDto -> {
+            LOGGER.info(String.format("Handling run request %s on TS ", taskDto.getTimestamp()));
             CoreValidRequest request = getCoreValidRequest(taskDto);
             coreValidClient.run(request);
         };
