+++ core-valid-adapter-app/src/main/resources/application.yml
@@ -1,0 +1,23 @@
+spring:
+  rabbitmq:
+    username: guest
+    password: guest
+    host: localhost
+    port: 5672
+  cloud:
+    stream:
+      default:
+        binder: rabbit
+        group: core-valid
+      bindings:
+        handleRun-in-0:
+          destination: run-task
+      rabbit:
+        default:
+          consumer:
+            auto-bind-dlq: true
+            max-attempts: 1
+
+core-valid-runner:
+  amqp:
+    queue-name: core-valid-requests
