+++ pom.xml
@@ -1,40 +1,31 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
+
 	<parent>
-		<groupId>org.springframework.boot</groupId>
-		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.6.1</version>
-		<relativePath/> <!-- lookup parent from repository -->
+		<groupId>com.farao-community.farao</groupId>
+		<artifactId>farao-parent-ws</artifactId>
+		<version>1.1.0</version>
 	</parent>
-	<groupId>com.farao-community.farao</groupId>
+
 	<artifactId>gridcapa-core-valid-adapter</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>gridcapa-core-valid-adapter</name>
 	<description>Application that turns a RabbitMQ request from a TaskDto to a CoreValidRequest</description>
-	<properties>
-		<java.version>11</java.version>
-		<gridcapa.core.valid.version>0.0.1-SNAPSHOT</gridcapa.core.valid.version>
-	</properties>
-	<dependencies>
-		<dependency>
-			<groupId>com.farao-community.farao</groupId>
-			<artifactId>gridcapa-core-valid-spring-boot-starter</artifactId>
-			<version>${gridcapa.core.valid.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-actuator</artifactId>
-		</dependency>
+	<packaging>pom</packaging>
 
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-test</artifactId>
-			<scope>test</scope>
-		</dependency>
-	</dependencies>
+	<licenses>
+		<license>
+			<name>Mozilla Public License, Version 2.0</name>
+			<url>https://www.mozilla.org/en-US/MPL/2.0/</url>
+		</license>
+	</licenses>
 
+	<modules>
+		<module>core-valid-adapter-app</module>
+	</modules>
 	<build>
 		<plugins>
 			<plugin>
@@ -43,5 +34,40 @@
 			</plugin>
 		</plugins>
 	</build>
+
+	<properties>
+		<maven.compiler.source>11</maven.compiler.source>
+		<maven.compiler.target>11</maven.compiler.target>
+		<spring-cloud.version>Hoxton.SR12</spring-cloud.version>
+		<task-manager.version>0.1.0</task-manager.version>
+		<core-valid-runner.version>0.0.1-SNAPSHOT</core-valid-runner.version>
+	</properties>
+
+	<dependencyManagement>
+		<dependencies>
+			<dependency>
+				<groupId>com.farao-community.farao</groupId>
+				<artifactId>gridcapa-task-manager-api</artifactId>
+				<version>${task-manager.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>com.farao-community.farao</groupId>
+				<artifactId>gridcapa-core-valid-spring-boot-starter</artifactId>
+				<version>${core-valid-runner.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>com.farao-community.farao</groupId>
+				<artifactId>gridcapa-core-valid-api</artifactId>
+				<version>${core-valid-runner.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>org.springframework.cloud</groupId>
+				<artifactId>spring-cloud-dependencies</artifactId>
+				<version>${spring-cloud.version}</version>
+				<type>pom</type>
+				<scope>import</scope>
+			</dependency>
+		</dependencies>
+	</dependencyManagement>
 
 </project>
