--- src/main/java/com/farao_community/farao/corevalid/adapter/CoreValidAdapterApplication.java
@@ -1,14 +1,0 @@
-package com.farao_community.farao.corevalid.adapter;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SuppressWarnings("hideutilityclassconstructor")
-@SpringBootApplication
-public class CoreValidAdapterApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CoreValidAdapterApplication.class, args);
-	}
-
-}
