--- src/main/resources/application.yml
@@ -1,14 +1,0 @@
-spring:
-  rabbitmq:
-    host: localhost
-    port: 5672
-  cloud:
-    stream:
-      bindings:
-        handleRun-in-0:
-          binder: rabbit
-          destination: core-valid-launch-task
-
-cse-cc-runner:
-  binding:
-    destination: core-valid-request
