--- src/test/java/com/farao_community/farao/corevalid/adapter/CoreValidAdapterApplicationTests.java
@@ -1,13 +1,0 @@
-package com.farao_community.farao.corevalid.adapter;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CoreValidAdapterApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
